user_pokemon_management/receiver/requirements.txt
uvicorn==0.27.0
connexion[flask]==3.0.5
swagger-ui-bundle==1.1.0
pykafka==2.8.0
requests==2.31.0
PyYAML==6.0.1
